Finishes initial welcome page using kotlin-styled
