+++ web/src/main/kotlin/orion/components/Logo.kt
@@ -1,0 +1,17 @@
+package logo
+
+import react.*
+import styled.*
+import kotlinx.css.*
+import js.utils.require
+
+val logoImg = require("../../../assets/images/logo/logo.png")
+
+fun RBuilder.logo() {
+    styledImg(alt = "Orion Logo", src = logoImg) {
+        css {
+            height = 40.px
+            padding(10.px)
+        }
+    }
+}
