--- web/src/main/kotlin/orion/components/logo/logo.css
@@ -1,3 +1,0 @@
-.Logo {
-    height: 60px;
-}
